mapping cora statuses
