+++ config/shortform_outputs_schema.toml
@@ -23,7 +23,7 @@ old_name = "formtype"
 Deduced_Data_Type = "category"
 
 [form_status]
-old_name = "status"
+old_name = "form_status"
 Deduced_Data_Type = "category"
 
 [ua_county]
