+++ src/developer_config.yaml
@@ -9,7 +9,7 @@ global:
   output_auto_outliers: False
   output_outlier_qa : False
   output_estimation_qa: False
-  output_short_form: False
+  output_short_form: True
   dev_test : False
   network_or_hdfs: network #whether to load from hdfs or network (local Python)
 years:
