+++ src/outputs/outputs_main.py
@@ -5,8 +5,6 @@ from datetime import datetime
 from typing import Callable, Dict, Any
 
 from src.outputs.short_form_out import run_shortform_prep, create_shortform_df
-from src.outputs.cora_mapping_temp_file_delete import create_cora_status_col
-from src.outputs.temp_file_to_be_deleted import combine_dataframes
 
 OutputMainLogger = logging.getLogger(__name__)
 
@@ -36,16 +34,11 @@ def run_output(
     NETWORK_OR_HDFS = config["global"]["network_or_hdfs"]
     output_path = config[f"{NETWORK_OR_HDFS}_paths"]["output_path"]
 
-    # Create combined ownership column using mapper
-    estimated_df = combine_dataframes(estimated_df, ultfoc_mapper)
 
-    # add cora status "form_status" using mapper
-    estimated_df = create_cora_status_col(estimated_df, cora_mapper)
+    # Prepare the shortform output dataframe
+    short_form_df = run_shortform_prep(estimated_df, cora_mapper, ultfoc_mapper, round_val=4)
 
-    # Creating blank columns for short form output
-    short_form_df = run_shortform_prep(estimated_df, round_val=4)
-
-    # Create short form output dataframe with required columns
+    # Create short form output dataframe with required columns from schema
     schema_path = "config/shortform_outputs_schema.toml"
     shortform_output = create_shortform_df(short_form_df, schema_path)
