--- src/outputs/temp_file_to_be_deleted.py
@@ -1,30 +1,0 @@
-import pandas as pd
-
-from src.utils.wrappers import exception_wrap
-
-
-@exception_wrap
-def combine_dataframes(main_df: pd.DataFrame, mapper_df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Combine two DataFrames using a left join based on specified columns.
-
-    Args:
-        main_df (pd.DataFrame): The main DataFrame.
-        mapper_csv_path (pd.DataFrame): The mapper DataFrame.
-
-    Returns:
-        pd.DataFrame: The combined DataFrame resulting from the left join.
-    """
-    try:
-        # Perform left join
-        combined_df = main_df.merge(
-            mapper_df, how="left", left_on="reference", right_on="ruref"
-        )
-        combined_df.drop(columns=["ruref"], inplace=True)
-
-        return combined_df
-
-    except Exception as e:
-        raise ValueError(
-            "An error occurred while combining main_df and mapper_df: " + str(e)
-        )
